pyproject.toml



[tool.poetry]
name = "learning_companion"
version = "0.1.0"
description = "An Agentinc Learning Companion grounded on RAG and WebSearchTools"
authors = ["adkaspar-google <adkaspar@google.com>"]
readme = "README.MD"
packages = [{include = "learning_companion"}]


[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.12"
tavily-python = "^0.5.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
pytest = "^8.3.4"
pypdf = "^5.1.0"
beautifulsoup4 = "^4.12.3"
langgraph = "^0.2.59"
langchainhub = "^0.1.21"
langchain-community = "^0.3.12"
google-api-python-client = "^2.155.0"
langchain-google-vertexai = "^2.0.9"
langchain-google-genai = "^2.0.7"
sphinx = "^8.1.3"
build = "^1.2.2.post1"
asyncio = "^3.4.3"
langchain-chroma = "^0.1.4"
langchain-google-community = "^2.0.3"
langgraph-checkpoint-sqlite = "^2.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.4"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
